% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_age.R
\name{bin_age}
\alias{bin_age}
\title{Binning AGE variable for IPUMS Microdata}
\usage{
bin_age(df, col)
}
\arguments{
\item{df}{dataframe containing an integer AGE variable}

\item{col}{name of the AGE variable}
}
\description{
Binning AGE variable for IPUMS Microdata
}
