% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fips_borough.R
\name{boro_mutate}
\alias{boro_mutate}
\title{Takes a fips county code and turns it into a borough name}
\usage{
boro_mutate(df, fips_col)
}
\arguments{
\item{df}{dataframe containing fips column}

\item{col}{fips column}
}
\description{
Takes a fips county code and turns it into a borough name
}
