% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fac_migrate1.R
\name{fac_migrate1}
\alias{fac_migrate1}
\title{Code MIGRATE1 IPUMS variable into a factor}
\usage{
fac_migrate1(df, var)
}
\arguments{
\item{df}{a dataframe containing MIGRATE1 column}

\item{var}{the MIGRATE1 column name}
}
\value{

}
\description{
Code MIGRATE1 IPUMS variable into a factor
}
