% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_ipums.R
\name{read_write_ipums}
\alias{read_write_ipums}
\title{Write IPUMS .dat files to database in chunks}
\usage{
read_write_ipums(ddi_path, dat_path, db_con, chunk_size, db_table_name)
}
\arguments{
\item{ddi_path}{path to xml file}

\item{dat_path}{path to .dat file}

\item{db_con}{connection object for database to write to}

\item{chunk_size}{how many records to read in at a time}

\item{db_table_name}{a string for what to name your table in db}
}
\description{
Write IPUMS .dat files to database in chunks
}
